hive - ITs will need fixing at some point in the future

+++ kafka-connect-aws-s3/src/main/scala/io/lenses/streamreactor/connect/aws/s3/formats/ParquetFormatWriter.scala
@@ -61,7 +61,7 @@ class ParquetFormatWriter(outputStreamFn: () => S3OutputStream) extends S3Format
 
     AvroParquetWriter
       .builder[AnyRef](outputFile)
-      .withRowGroupSize(DEFAULT_BLOCK_SIZE.toLong)
+      .withRowGroupSize(DEFAULT_BLOCK_SIZE)
       .withPageSize(DEFAULT_PAGE_SIZE)
       .withSchema(schema)
       .build()
